Update state machine factory
